pyproject.toml
[tool.poetry]
name = "mptt"
version = "0.1.0"
description = "learning tree stuff with mptt dropbox style clone"
authors = ["Enrique Galindo <egalindo@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.4"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
